+++ CHANGELOG.md
@@ -14,9 +14,7 @@
 
 ### Updates
 
-* changed coloring.
-* binary format now used. An extra mask cancels/hides the zeros
-* string replace() removed to increase performance.
+* bitmode added for retro feel (up to 50 % reduced execution time).
 
 ### Bug Fixes